+++ openapi-generator/example/example.dart
@@ -1,9 +1,9 @@
 import 'package:openapi_generator_annotations/openapi_generator_annotations.dart';
 
 @Openapi(
-    additionalProperties:
-        AdditionalProperties(pubName: 'petstore_api', pubAuthor: 'Johnny dep'),
-    inputSpecFile: 'example/openapi-spec.yaml',
-    generatorName: Generator.dart,
+    additionalProperties: AdditionalProperties(
+        pubName: 'petstore_api', pubAuthor: 'Johnny dep...'),
+    inputSpecFile: 'https://google.com',
+    generatorName: Generator.dio,
     outputDirectory: 'api/petstore_api')
 class Example extends OpenapiGeneratorConfig {}