+++ openapi-generator/test/config.dart
@@ -1,0 +1,9 @@
+import 'package:openapi_generator_annotations/openapi_generator_annotations.dart';
+
+@Openapi(
+    additionalProperties:
+        AdditionalProperties(pubName: 'petstore_api', pubAuthor: 'Johnny dep'),
+    inputSpecFile: 'example/openapi-spec.yaml',
+    generatorName: Generator.dart,
+    outputDirectory: 'api/petstore_api')
+class Config {}